Updated Map
I started updating the dorm maps. The main halls all exist on one map, and all of the rooms exist in their own map. When entering a room, it diplays the room name once. It will do this every time you enter the room. Takes a bit more work when you have multiple different room names all in the map, but I made it work. I may download an extension that allows the player to continue moving when a text box is showing, because right now, the player is frozen for like 1.25 second. Not a huge deal, but it would be nice to not have that. An event also calls on some javaScript code, to reset self switches when exiting. This is to ensure that the message on the room name only displays once, but also appears every time you enter the room.