Created the first Scenario
I created the first scenario. It is not 100% seamless, but you bascially interact with the person, you choose an option, and it gives dialogue based off of your choice. The scene then fades to black (For time passing, we may need a sound effect), then fades back in. (In one instance, the girl is left alone. During this time, she will be moved off screen.) I also made it so the scene won't appear unless a switch is turned on. This is long term planning, so when we are dynamically loading scenes, we can effectivley hide all of the scenes until they are available. It also turns on a self switch that will prevent the scene from becoming available again. This might be a good practice, as it will prevent you from re-doing a scene. I would rather not load a scene that you have already done than re-do scenes. (Both are bad).