Added event 9
I added in the scripting for event 9 (Helping an international student).
I also added a small side room for the class room. It is more placeholder than anything since I accidentally deleted the side rooms for the tall building. We were going to change things about that room anyway so I am not really worried about it.