Added in Excel to Json functionality
So I added in functionality for parsing from excel to JSON. It puts everything as one line, but I can later convert from one line to each separate file for easier management with the items that use it. I think as long as everything is ordered and incremental, we should be good. I did have to install some packages to get this to work though... like over 1900... (I installed like two packages, that had that many files don't yell)