Got Meta-Data working fully
Good news: This is now working fully. I even updated the excel sheet so it prevents you from entering duplicate data under title and ID. Now, the code will automatically update the created json file so it is just an array. Then I iterate over the entire array, creating a file for each scenario. The file name is based off of the ID (which is why I forced it to be unique). So Event 5 will have a file name "Scenario5.json". I left the old version in though, so if we wanted to use the array notation and just have one large file, we can.