Half-Populating tables
The framework for populating the tables is now present. Under the PopulateTables function, the system is able to read in the json file from the system. There is code that proves that is an array of json objects. Vincent will be working on using that framework for populating and then using all of the tables in RPG maker.