Updated scenarios
I have updated over half of my scenarios so they now output consequences (or information about your consequences) to the social media handler. This is done by having the main scenario call on a common event, which will then use the data used in the variables to determine what to say. It should be noted that while this process could have been done entirely on the main event, those were starting to get really crammed with things, so this was to help separate those concerns out.