Started Pre-Message (example on event 3)
Scenario 3 (top most left scenario) has a stand-in pre-scenario message. It requires a slight change in implementation. The NCP that flips a switch now turns on switch 203. Once that happens, common event 103 (labaled scene 3 startup) will log things to the social media handler (For now, it just gives the row and column on the grid which is 1, 1). It then disables the logger, turns off switch 203 (as to prevent the event from looping), THEN turns on switch 3 to enable the scene. 

NOTE FOR VINCENT: Your algorithm will need a TINY adjustment. Rather than indexing directly to to the switch (meaning if the index is 3, you enable switch 3), you will have to add 200 to that. So now it will turn on switch 203. The logic I am employing should handle the rest. Play though it for an example. Will be doing more of these tonight, just wanted to get this put out there.