Added in number of correct scenes completed
I added in a change so now the game keeps track of the number of correct scenarios completed. This is under var 500. As for when to initialize the end-game event, I would switch a switch to true once Vincen't s thing realizes that there are no more scenarios left. Then I would have a common event run when that switch gets turned on. Transport the player to a blankish map with Dr. H-R with an event set to autorun and boom we good.