Add in stand in end-game message; add in baloon icons
I added in a message about Dr. H-R saying how you did. It is a stand in, but it properly calls on the variable 500 which we will be updating when the correct option is working. 
I also add in bubble icons to all of my scenarios: Meaning when you choose the correct response, you have a little heart appear once you are finished.